rename pkg, fix import, rename files, add go mod
