+++ tgb/rule/base_rule.go
@@ -1,4 +1,4 @@
-package rules
+package rule
 
 import (
 	"fmt"
