+++ tgb/rule/min.go
@@ -1,4 +1,4 @@
-package rules
+package rule
 
 import (
 	"fmt"
