+++ tgb/rule/required.go
@@ -1,4 +1,4 @@
-package rules
+package rule
 
 import (
 	"reflect"
